% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-bibentry.R
\encoding{UTF-8}
\name{as_bibentry}
\alias{as_bibentry}
\alias{as_bibentry.default}
\alias{as_bibentry.character}
\alias{as_bibentry.NULL}
\alias{as_bibentry.list}
\alias{as_bibentry.cff}
\alias{as_bibentry.cff_ref_lst}
\alias{as_bibentry.cff_ref}
\title{Create \code{\link{bibentry}} objects from multiple sources}
\usage{
as_bibentry(x, ...)

\method{as_bibentry}{default}(x, ...)

\method{as_bibentry}{character}(x, ..., what = c("preferred", "references", "all"))

\method{as_bibentry}{`NULL`}(x, ...)

\method{as_bibentry}{list}(x, ...)

\method{as_bibentry}{cff}(x, ..., what = c("preferred", "references", "all"))

\method{as_bibentry}{cff_ref_lst}(x, ...)

\method{as_bibentry}{cff_ref}(x, ...)
}
\arguments{
\item{x}{The source used to generate the \code{bibentry} object with
\CRANpkg{cffr}. It can be:
\itemize{
\item A missing or \code{NULL} value, which retrieves the \code{DESCRIPTION} file from
your in-development package.
\item An existing \code{cff} object created with \code{\link[=cff]{cff()}}, \code{\link[=cff_create]{cff_create()}} or
\code{\link[=as_cff]{as_cff()}}.
\item A path to a \code{CITATION.cff} file (\code{"CITATION.cff"}).
\item The name of an installed package (\code{"jsonlite"}).
\item A path to a \code{DESCRIPTION} file (\code{"DESCRIPTION"}).
}}

\item{...}{Additional arguments passed to or from methods.}

\item{what}{Fields to extract from a full \code{cff} object. It can be:
\itemize{
\item \code{preferred}: Create a single entry with the main citation information
of the package (key \code{preferred-citation}).
\item \code{references}: Extract all entries of the \code{references} key.
\item \code{all}: Extract both the \code{preferred-citation} and \code{references} keys.
}

See \code{vignette("r-cff", package = "cffr")}.}
}
\value{
A \code{\link[utils:bibentry]{bibentry}} object with one or more entries.
}
\description{
This function creates \code{\link{bibentry}} objects from multiple metadata sources
(\code{\link{cff}} objects, \code{DESCRIPTION} files and other sources). The inverse
transformation from a \code{bibentry} object to \code{\link{cff_ref_lst}} can be performed
with the corresponding \code{\link[=as_cff.bibentry]{as_cff.bibentry()}} method.

With \code{\link[=toBibtex.cff]{toBibtex()}}, you can convert \code{\link{cff}} objects to
BibTeX markup on the fly. See \strong{Examples}.
}
\details{
An \R \code{bibentry} object is the representation of a BibTeX entry. These
objects can be converted to BibTeX markup with \code{\link[utils:toBibtex]{utils::toBibtex()}}, which
creates an object of class \code{Bibtex} that can be printed and exported as a
valid BibTeX entry.

\code{as_bibentry()} tries to map the information of the source \code{x} into a \code{\link{cff}}
object and maps the metadata to BibTeX as described in
\code{vignette("bibtex-cff", package = "cffr")}.
}
\examples{
\donttest{
# From a `cff` object ----
cff_object <- cff()

cff_object

# A `bibentry` object.
bib <- as_bibentry(cff_object)

class(bib)

bib

# Print as BibTeX.
toBibtex(bib)

# The S3 method also supports this.
toBibtex(cff_object)

# Other sources ----
# From a CITATION.cff file.

path <- system.file("examples/CITATION_complete.cff", package = "cffr")
cff_file <- as_bibentry(path)

cff_file

# For an installed package with options.
installed_package <- as_bibentry("jsonvalidate", what = "all")

installed_package

# Use a DESCRIPTION file.
path2 <- system.file("examples/DESCRIPTION_gitlab", package = "cffr")
desc_file <- as_bibentry(path2)

toBibtex(desc_file)
}

}
\references{
Patashnik O (1988). "BIBTEXing."
\url{https://osl.ugr.es/CTAN/biblio/bibtex/base/btxdoc.pdf}.

Haines R, The Ruby Citation File Format Developers (2022).
"Ruby CFF Library." \doi{10.5281/zenodo.7294987}.
\url{https://github.com/citation-file-format/ruby-cff}.

Hernangómez D (2022). "BibTeX and CFF, a potential crosswalk."
\CRANpkg{cffr} vignette.
\url{https://docs.ropensci.org/cffr/articles/bibtex-cff.html}.
}
\seealso{
\itemize{
\item \code{\link[utils:bibentry]{utils::bibentry()}} documents the \code{bibentry} class.
\item \code{vignette("r-cff", package = "cffr")} explains how the
metadata of a package is mapped to produce a \code{cff} object.
\item \code{vignette("bibtex-cff", package = "cffr")} provides details about the
internal mapping performed between \code{cff} objects and BibTeX
markup, both \code{cff} to BibTeX and BibTeX to \code{cff}.
\item \code{\link[utils:toBibtex]{utils::toBibtex()}} converts \code{bibentry} objects to BibTeX markup.
}

Work with BibTeX metadata:
\code{\link[=cff_read]{cff_read()}},
\code{\link[=cff_read_bib_text]{cff_read_bib_text()}},
\code{\link[=cff_write_bib]{cff_write_bib()}},
\code{\link[=encoded_utf_to_latex]{encoded_utf_to_latex()}}

Convert between \R classes:
\code{\link[=as_cff]{as_cff()}},
\code{\link[=as_cff_person]{as_cff_person()}},
\code{\link{cff_class}}
}
\concept{bibtex}
\concept{conversions}
