% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{bibtex}
\alias{bibtex}
\alias{cff_read_bib}
\title{BibTeX utils}
\usage{
cff_read_bib(file)
}
\arguments{
\item{file}{The path to a \code{.bib} file.}
}
\value{
\code{\link[=cff_read_bib]{cff_read_bib()}} returns a \verb{citation, bibentry} object as produced by
\code{\link[=citation]{citation()}}
}
\description{
Functions for reading and creating \code{.bib} files.
}
\details{
These are simple I/O functions for working with BibTeX files. Other options
are:
\itemize{
\item \href{https://github.com/ropensci/bibtex}{bibtex} package.
\item \href{https://github.com/ropensci/RefManageR}{RefManageR} package.
\item \code{\link[knitr:write_bib]{knitr::write_bib()}}.
}
}
\note{
\strong{R} does not have full support for BibTex. Specifically, \strong{InBook} has
limited support and \strong{Conference} is not supported. The latter would be
parsed as \strong{InProceedings} (Patashnik (1988)).

See \code{\link[=bibentry]{bibentry()}} for more info.
}
\examples{
# Reading .bib files

file <- system.file("examples/bibtex_example.bib",
  package = "cffr"
)
parsed <- cff_read_bib(file)


print(parsed, bibtex = TRUE)

# Compare with the file
cat(readLines(file), sep = "\n")

# To cffr citation

cff_parse_citation(parsed)
}
\references{
Patashnik, O. (1988, February). \emph{BIBTEXTING}. BibTeX - Process
Bibliographies for LATEX, Etc. Retrieved December 1, 2021, from
\url{https://osl.ugr.es/CTAN/biblio/bibtex/base/btxdoc.pdf}.
}
\seealso{
\code{\link[=citation]{citation()}}, \code{\link[=bibentry]{bibentry()}}, \code{\link[knitr:write_bib]{knitr::write_bib()}}.

Other bibtex: 
\code{\link{encoded_utf_to_latex}()}
}
\concept{bibtex}
