% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_parse_person.R
\name{cff_authors_to_person}
\alias{cff_authors_to_person}
\title{Convert a \code{cff} \code{authors} key to an R \code{person} object}
\usage{
cff_authors_to_person(cff_authors)
}
\arguments{
\item{cff_authors}{A list of authors from a \code{cff} object, either
generated with \code{\link[=cff_parse_person]{cff_parse_person()}} or \code{\link[=cff_create]{cff_create()}}.}
}
\value{
A \code{person} object
}
\description{
Parse an authors key of a \code{cff} object to an R \code{\link[utils:person]{utils::person()}} object.
}
\examples{

rmark <- cff_create("rmarkdown")

rmark$authors

cff_authors_to_person(rmark$authors)


# From cff_parse_person

cff_aut <- cff_parse_person("Julio Iglesias <fake@email.com>")
cff_authors_to_person(cff_aut)
}
\seealso{
\code{\link[utils:person]{utils::person()}}

Other parsers: 
\code{\link{cff_parse_citation}()},
\code{\link{cff_parse_person}()}
}
\concept{parsers}
