% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff-create.R
\encoding{UTF-8}
\name{cff_create}
\alias{cff_create}
\title{Create a \code{\link{cff}} object from multiple sources}
\usage{
cff_create(
  x,
  keys = list(),
  cff_version = "1.2.0",
  gh_keywords = TRUE,
  dependencies = TRUE,
  authors_roles = c("aut", "cre")
)
}
\arguments{
\item{x}{The source used to generate the \code{\link{cff}} object. It can be:
\itemize{
\item A missing value, which retrieves the \code{DESCRIPTION} file from your
in-development \R package.
\item An existing \code{\link{cff}} object.
\item The name of an installed package (\code{"jsonlite"}).
\item A path to a \code{DESCRIPTION} file (\code{"./DESCRIPTION"}).
}}

\item{keys}{A list of additional keys to add to the \code{\link{cff}} object. See
\code{\link[=cff_modify]{cff_modify()}}.}

\item{cff_version}{The Citation File Format schema version used for the
generated metadata.}

\item{gh_keywords}{A logical value. If \code{TRUE} and the package is hosted on
GitHub, add the repository topics as keywords.}

\item{dependencies}{A logical value. If \code{TRUE}, add installed package
dependencies to the \code{references} CFF key. See \strong{Details}.}

\item{authors_roles}{Roles to be considered as authors of the package when
generating the \code{CITATION.cff} file. See \strong{Details}.}
}
\value{
A \code{\link{cff}} object.
}
\description{
Create a full, possibly valid \code{\link{cff}} object from a given source. This
object can be written to a \code{CITATION.cff} file with \code{\link[=cff_write]{cff_write()}}. See
\strong{Examples}.

This function performs most metadata extraction and conversion in
\CRANpkg{cffr}.
}
\details{
If \code{x} is a path to a \code{DESCRIPTION} file or if \code{inst/CITATION} is not
present in your package, \CRANpkg{cffr} auto-generates a
\code{preferred-citation} key using the information provided in that file.

When \code{inst/CITATION} is present, every field from the package's
\code{DESCRIPTION}, including custom fields, is made available to that file
through its \code{meta} object. If the file cannot be evaluated with those
metadata, \code{cff_create()} does not retry with metadata from another package.

When \code{dependencies = TRUE}, \code{cff_create()} starts from the first reference
returned by \code{\link[utils:citation]{utils::citation()}} for each installed dependency. Citation
fields such as title, abstract, authors and year are preserved. Each
reference is adapted to the CFF software-reference type and enriched with
the dependency scope and version constraint from the source \code{DESCRIPTION},
URLs and repository information from the installed dependency's
\code{DESCRIPTION} and the canonical CRAN DOI only when the dependency is
available from CRAN. A release date, when present, determines the year, with
the citation year retained otherwise.

By default, only persons whose role in the \code{DESCRIPTION} file of the package
is author (\code{"aut"}) or maintainer (\code{"cre"}) are considered package authors.
The default setting can be controlled with the \code{authors_roles} argument. See
\strong{Details} on \code{\link[utils:person]{utils::person()}} for additional information about person
roles.
}
\examples{
\donttest{
# Installed package.
cff_create("jsonlite")

# Demo file.
demo_file <- system.file("examples/DESCRIPTION_basic", package = "cffr")
cff_create(demo_file)

# Add additional keys.

newkeys <- list(
  message = "This overwrites keys",
  abstract = "New abstract",
  keywords = c("A", "new", "list", "of", "keywords"),
  authors = as_cff_person("New author")
)

cff_create(demo_file, keys = newkeys)

# Update a key in a list, such as authors or contacts.
# Add a new contact.

old <- cff_create(demo_file)

new_contact <- append(
  old$contact,
  as_cff_person(person(
    given = "I am",
    family = "New Contact"
  ))
)

cff_create(demo_file, keys = list("contact" = new_contact))
}
}
\seealso{
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema guide}.
\itemize{
\item \code{vignette("cffr", package = "cffr")} shows an introduction to
manipulating \code{\link{cff}} objects.
\item \code{vignette("r-cff", package = "cffr")} provides details about how the
metadata of a package is mapped to produce a \code{cff} object.
}

Core \CRANpkg{cffr} workflow:
\code{\link[=cff]{cff()}},
\code{\link[=cff_modify]{cff_modify()}},
\code{\link[=cff_validate]{cff_validate()}},
\code{\link[=cff_write]{cff_write()}}
}
\concept{core}
