% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_create_cff_person.R, R/cff-methods.R
\name{cff_create_cff_person}
\alias{cff_create_cff_person}
\alias{as.person.cff}
\title{Create a person with the corresponding \code{\link[=cff-class]{cff}} structure}
\usage{
cff_create_cff_person(person)

\method{as.person}{cff}(x)
}
\arguments{
\item{person}{It can be either:
\itemize{
\item A \code{person} or list of \code{person} object created with \code{\link[utils:person]{utils::person()}}.
\item A \code{character} object or vector representing a person or persons.
See \strong{Examples}.
}}

\item{x}{\code{cff} object representing a person or entity.}
}
\value{
\code{cff_create_cff_person()} returns A list of persons or entities with class
\code{cff} converted to the \href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema}.

\code{as.person.cff()} returns a \code{person} object.
}
\description{
Create a \code{person} or \code{entity} as defined by the
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema}.

\code{\link[=cff_create_cff_person]{cff_create_cff_person()}} can convert the following objects:
\itemize{
\item Objects with class \code{person} as provided by \code{\link[utils:person]{utils::person()}}.
\item A \code{character} string with the definition of an author or several authors,
using the standard BibTeX notation. See Markey (2007) for a full
explanation.
}

Additionally, it is also provided a method for \code{\link[=as.person]{as.person()}}, that can
convert \code{\link[=cff-class]{cff}} objects to \code{person} objects as
provided by \code{\link[utils:person]{utils::person()}}.

This method only works with CFF persons, not with full \code{cff} objects.
}
\details{
\code{cff_create_cff_person()} uses a custom algorithm that tries to break a name
as explained in Section 11 of "Tame the BeaST" (Markey, 2007):
\itemize{
\item \verb{First von Last}
\item \verb{von Last, First}
\item \verb{von Last, Jr, First}
}

\code{First} is mapped to the CFF field \code{given-names}, \code{von} to \code{name-particle},
\code{Last} to \code{family-names} and \code{Jr} to \code{name-suffix}.

In the case of entities, the whole \code{character} would be mapped to \code{name}.
It is a good practice to "protect" entity's names with \code{{}}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Don't do
entity <- "Elephant and Castle"
cff_create_cff_person(entity)
#> - name: Elephant
#> - name: Castle

# Do
entity_protect <- "\{Elephant and Castle\}"
cff_create_cff_person(entity_protect)
#> - name: Elephant and Castle
}\if{html}{\out{</div>}}

\code{cff_create_cff_person()} would try to add as many information as possible.
On \code{character} string coming from \code{\link[utils:person]{format(person())}} the
email and the ORCID would be gathered as well.
}
\examples{
# Create a person object
a_person <- person(
  given = "First", family = "Author",
  role = c("aut", "cre"),
  email = "first.last@example.com", comment = c(
    ORCID = "0000-0001-8457-4658",
    affiliation = "An affiliation"
  )
)

a_person

cff_person <- cff_create_cff_person(a_person)

cff_person

# Back to person object with S3 Method
as.person(cff_person)

# Parse a string
a_str <- paste0(
  "Julio Iglesias <fake@email.com> ",
  "(<https://orcid.org/0000-0001-8457-4658>)"
)
cff_create_cff_person(a_str)

# Several persons
persons <- c(person("Clark", "Kent"), person("Lois", "Lane"))

cff_create_cff_person(persons)

# Or you can use BibTeX style if you prefer

x <- "Frank Sinatra and Dean Martin and Davis, Jr., Sammy and Joey Bishop"

cff_create_cff_person(x)

cff_create_cff_person("Herbert von Karajan")
}
\references{
\itemize{
\item Patashnik, Oren. "BIBTEXTING" February 1988.
\url{https://osl.ugr.es/CTAN/biblio/bibtex/base/btxdoc.pdf}.
\item Markey, Nicolas. "Tame the BeaST"
\emph{The B to X of BibTeX, Version 1.4} (October 2007).
\url{https://osl.ugr.es/CTAN/info/bibtex/tamethebeast/ttb_en.pdf}.
}

See \strong{Examples} for more information.
}
\seealso{
Examples in \code{vignette("cffr", "cffr")} and \code{\link[utils:person]{utils::person()}}.

\code{\link[utils:person]{utils::person()}}

Other functions for converting between \strong{R} classes:
\code{\link{as_bibentry}()}

Other S3 Methods for \code{cff}:
\code{\link{as_bibentry}()}
}
\concept{coercing}
\concept{s3method}
