% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_create_person.R
\name{cff_create_person}
\alias{cff_create_person}
\title{Create a person with the corresponding \code{\link[=cff-class]{cff}} structure}
\usage{
cff_create_person(person, hint = c("auto", "person", "bibtex"))
}
\arguments{
\item{person}{It can be either:
\itemize{
\item A \code{person} or list of \code{person} object created with \code{\link[utils:person]{utils::person()}}.
\item A \code{character} object representing a person or persons.
See \strong{Examples}.
}}

\item{hint}{Either \code{auto}, \code{person} or \code{bibtex}. See \strong{Details}.}
}
\value{
A list of persons or entities with class \code{cff} converted to the \href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema}.
}
\description{
Create a \code{person} or \code{entity} as defined by the
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema}.

\code{\link[=cff_create_person]{cff_create_person()}} can convert the following objects:
\itemize{
\item Objects with class \code{person} as provided by \code{\link[utils:person]{utils::person()}}.
\item A \code{character} string with the definition of an author or several authors,
using the standard BibTeX notation. See Markey (2007) for a full
explanation.
}
}
\details{
\code{cff_create_person()} would try to guess what type of object has been
provided in \code{person}. Possible options are:
\enumerate{
\item If \code{person} is a \code{\link[utils:person]{person}} object it would be parsed
with a specific algorithm.
\item However, if \code{person} is a \code{character} there are two options:
\enumerate{
\item The \code{character} be a printed version of \code{\link[utils:person]{utils::person()}}. In this
case it may include information of email, ORCID, etc. See
\code{\link[utils:person]{utils::as.person()}} and \strong{Details} in that same help page.
\item It can be just a person with BibTeX notation such as
\code{"von Doe, John and Smith, Jane"}.
}
}

In \strong{case 1} (\code{person = "person"}) \code{hint = "auto"} wouldn't affect and
internally the algorithm used with \code{hint = "person"} would be always used.

\code{hint} argument would affect \strong{case 2} (i.e. \code{person = "character"}). If
\code{hint = "auto"} \code{cff_create_person()} would try to guess the best
alternative, however it may be incorrect, so it is possible to specify the
conversion algorithm:
\itemize{
\item \code{hint = "person"} would try to convert the \code{character} to \code{person} object
with \code{\link[utils:person]{utils::as.person()}} and then proceed with the algorithm for
\code{person} objects.
\item \code{hint = "bibtex"} would use a special algorithm to create the \code{cff}
person. This option is better for strings like \code{"von Last, Jr, First"}
and produces improved mappings to CFF keys as \code{name-particle} and
\code{name-suffix}. It is also aware of protected names such as
\verb{\{John Doe and Jane Smith inc.\}}.
}
}
\examples{
# Create a person object
a_person <- person(
  given = "First", family = "Author",
  role = c("aut", "cre"),
  email = "first.last@example.com", comment = c(
    ORCID = "0000-0001-8457-4658",
    affiliation = "An affiliation"
  )
)
cff_create_person(a_person)

# Parse a string
a_str <- paste0(
  "Julio Iglesias <fake@email.com> ",
  "(<https://orcid.org/0000-0001-8457-4658>)"
)
cff_create_person(a_str)

# Several persons
persons <- c(person("Clark", "Kent"), person("Lois", "Lane"))

cff_create_person(persons)

# Or you can use BibTeX style if you prefer

x <- "Frank Sinatra and Dean Martin and Davis, Jr., Sammy and Joey Bishop"

cff_create_person(x)

cff_create_person("Herbert von Karajan")
}
\references{
\itemize{
\item Patashnik, Oren. "BIBTEXTING" February 1988.
\url{https://osl.ugr.es/CTAN/biblio/bibtex/base/btxdoc.pdf}.
\item Markey, Nicolas. "Tame the BeaST."
\emph{The B to X of BibTeX, Version 1.4} (October 2007).
\url{https://osl.ugr.es/CTAN/info/bibtex/tamethebeast/ttb_en.pdf}.
}

See \strong{Examples} for more information.
}
\seealso{
Examples in \code{vignette("cffr", "cffr")} and \code{\link[utils:person]{utils::person()}}.
}
\concept{coercing}
