% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_description.R
\name{cff_description}
\alias{cff_description}
\title{Convert \code{DESCRIPTION} to cff format}
\usage{
cff_description(
  desc_path = "DESCRIPTION",
  message = "If you use this software, please cite it using these metadata."
)
}
\arguments{
\item{desc_path}{Path to a \code{DESCRIPTION} file}

\item{message}{A message to the human reader of the \code{CITATION.cff} file to
let them know what to do with the citation metadata.}
}
\value{
A character value ready for writing into \code{yaml} format.
}
\description{
Extract information from a \code{DESCRIPTION} file and creates an object
to be converted into a \code{CITATION.cff} file.
}
\seealso{
\code{\link[yaml:as.yaml]{yaml::as.yaml()}}

\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Guide to Citation File Format schema version 1.2.0}
}
\keyword{internal}
