% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff-gha-update.R
\encoding{UTF-8}
\name{cff_gha_update}
\alias{cff_gha_update}
\title{Install a \CRANpkg{cffr} GitHub Actions workflow}
\usage{
cff_gha_update(path = ".", overwrite = FALSE)
}
\arguments{
\item{path}{Project root directory.}

\item{overwrite}{A logical value. If \code{TRUE}, overwrite an existing workflow.}
}
\value{
Invisibly returns \code{NULL}. This function is called for its side effects.
}
\description{
This function installs a
\href{https://github.com/features/actions}{GitHub Actions} workflow in your
repository. The workflow
updates your \code{CITATION.cff} file when any of these events occur:
\itemize{
\item You publish a new release of the package.
\item Your \code{DESCRIPTION} or \code{inst/CITATION} file is modified.
\item You run the workflow manually.
}
}
\details{
Workflow triggers can be modified. See \href{https://docs.github.com/en/actions/learn-github-actions/events-that-trigger-workflows}{Events that trigger workflows}.
}
\examples{
\dontrun{
cff_gha_update()
}
}
\seealso{
Keep \file{CITATION.cff} up to date:
\code{\link{cff_git_hook}}
}
\concept{git}
