% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_read_biblines.R
\name{cff_read_biblines}
\alias{cff_read_biblines}
\title{Create a \code{\link[=cff-class]{cff}} object from BibTeX lines}
\usage{
cff_read_biblines(x, encoding = "UTF-8", ...)
}
\arguments{
\item{x}{A vector of characters with the full BibTeX string.}

\item{encoding}{Encoding to be assumed for \code{x}, see \code{\link[=readLines]{readLines()}}.}

\item{...}{
  Arguments passed on to \code{\link[=cff_read_bib]{cff_read_bib}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A \code{\link[=cff-class]{cff}} object ready to be used with other functions (i.e.
\code{\link[=cff_create]{cff_create()}}.
}
\description{
Convert a \code{\link[=character]{character}} representing a BibTeX entry to a
\code{\link[=cff-class]{cff}} object.
}
\details{
This is a helper function that writes \code{x} to a \verb{*.bib} file and reads it with
\code{\link[=cff_read_bib]{cff_read_bib()}}.

This function requires \CRANpkg{bibtex} (>= 0.5.0) and uses
\code{\link[bibtex:read.bib]{bibtex::read.bib()}}.
}
\examples{
if (requireNamespace("bibtex", quietly = TRUE)) {
  x <- c(
    "@book{einstein1921,
      title        = {Relativity: The Special and the General Theory},
      author       = {Einstein, Albert},
      year         = 1920,
      publisher    = {Henry Holt and Company},
      address      = {London, United Kingdom},
      isbn         = 9781587340925
  }",
    "@misc{misc-full,
      title        = {Handing out random pamphlets in airports},
      author       = {Joe-Bob Missilany},
      year         = 1984,
      month        = oct,
      note         = {This is a full MISC entry},
      howpublished = {Handed out at O'Hare}
  }"
  )


  cff_read_biblines(x)
}
}
\seealso{
\code{\link[=cff_read_bib]{cff_read_bib()}} for reading \verb{*.bib} files.

Other functions for working with BibTeX format:
\code{\link{cff_read}()},
\code{\link{cff_to_bibentry}()},
\code{\link{cff_write_bib}()},
\code{\link{encoded_utf_to_latex}()}

Reading other files as \code{cff}:
\code{\link{cff_read}()}
}
\concept{bibtex}
\concept{reading}
