% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-schema.R
\name{cff_schema_keys}
\alias{cff_schema_keys}
\title{Valid keys in a \code{CITATION.cff} file.}
\source{
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema version 1.2.0}
}
\usage{
cff_schema_keys(sorted = FALSE)
}
\arguments{
\item{sorted}{Logical \code{TRUE/FALSE}. Should the keys be arranged
alphabetically?}
}
\value{
A vector of characters with the names of the valid keys to be used on a
Citation File Format version 1.2.0
}
\description{
Set of valid keys as specified on the
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema version 1.2.0}
}
\examples{

cff_schema_keys(sorted = TRUE)
}
