% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff-write.R
\encoding{UTF-8}
\name{cff_write}
\alias{cff_write}
\title{Write a \code{CITATION.cff} file}
\usage{
cff_write(
  x,
  outfile = "CITATION.cff",
  keys = list(),
  cff_version = "1.2.0",
  gh_keywords = TRUE,
  r_citation = FALSE,
  dependencies = TRUE,
  validate = TRUE,
  verbose = TRUE,
  authors_roles = c("aut", "cre"),
  encoding = "UTF-8"
)
}
\arguments{
\item{x}{The source used to generate the \code{\link{cff}} object. It can be:
\itemize{
\item A missing value, which retrieves the \code{DESCRIPTION} file from your
in-development \R package.
\item An existing \code{\link{cff}} object.
\item The name of an installed package (\code{"jsonlite"}).
\item A path to a \code{DESCRIPTION} file (\code{"./DESCRIPTION"}).
}}

\item{outfile}{The name and path of the \code{CITATION.cff} to be created.
Relative paths are resolved from the current working directory,
independently of the source specified in \code{x}.}

\item{keys}{A list of additional keys to add to the \code{\link{cff}} object. See
\code{\link[=cff_modify]{cff_modify()}}.}

\item{cff_version}{The Citation File Format schema version used for the
generated metadata.}

\item{gh_keywords}{A logical value. If \code{TRUE} and the package is hosted on
GitHub, add the repository topics as keywords.}

\item{r_citation}{A logical value. If \code{TRUE}, the \R package
citation (\code{inst/CITATION}) is created or updated relative to the current
working directory. \strong{No backup copy is created}. For more control, use
\code{\link[=cff_write_citation]{cff_write_citation()}}.}

\item{dependencies}{A logical value. If \code{TRUE}, add installed package
dependencies to the \code{references} CFF key. See \strong{Details}.}

\item{validate}{A logical value. If \code{TRUE}, validate the new file with
\code{\link[=cff_validate]{cff_validate()}}.}

\item{verbose}{A logical value. If \code{TRUE}, the function displays
informative messages.}

\item{authors_roles}{Roles to be considered as authors of the package when
generating the \code{CITATION.cff} file. See \strong{Details}.}

\item{encoding}{The name of the encoding to be assumed. Default is \code{"UTF-8"},
but it can be any other value accepted by \code{\link[base:iconv]{base::iconv()}}, such as
\code{"ASCII//TRANSLIT"}.}
}
\value{
Invisibly returns the generated \code{\link{cff}} object. This function is called
primarily for its side effect of writing a \code{CITATION.cff} file.
}
\description{
\code{cff_write()} is the primary workflow for package development.

This function writes a \code{CITATION.cff} file for a given package. It wraps
\code{\link[=cff_create]{cff_create()}} to create the \code{\link{cff}} object, then writes it to a
YAML-formatted file in one command.
}
\details{
For details of \code{authors_roles} and dependency extraction, see \code{\link[=cff_create]{cff_create()}}.

The \code{x} argument identifies the metadata source. It does not determine the
output directory. This allows you to create a \code{CITATION.cff} file from a
package or file located outside the current working directory.

When creating and writing a \code{CITATION.cff} file for a package in the current
working directory, this function adds the pattern \code{"^CITATION\\.cff$"} to
the local \code{.Rbuildignore} file.
}
\examples{
\donttest{
tmpfile <- tempfile(fileext = ".cff")
cff_obj <- cff_write("jsonlite", outfile = tmpfile)

cff_obj

# Force cleanup.
file.remove(tmpfile)
}
}
\seealso{
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema guide}.

Core \CRANpkg{cffr} workflow:
\code{\link[=cff]{cff()}},
\code{\link[=cff_create]{cff_create()}},
\code{\link[=cff_modify]{cff_modify()}},
\code{\link[=cff_validate]{cff_validate()}}

Write citation metadata files:
\code{\link[=cff_write_bib]{cff_write_bib()}}
}
\concept{core}
\concept{writing}
