% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cffr-package.R
\docType{package}
\encoding{UTF-8}
\name{cffr-package}
\alias{cffr}
\alias{cffr-package}
\title{cffr: Generate Citation File Format ('CFF') Metadata}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Citation File Format ('CFF') version 1.2.0 \doi{10.5281/zenodo.5171937} is a human- and machine-readable file format for software citation metadata. Core utilities generate, read, write and validate Citation File Format metadata for 'R' packages.
}
\references{
Hernangómez D (2021). "cffr: Generate Citation File Format Metadata for R
Packages." \emph{Journal of Open Source Software}, \strong{6}(67), 3900.
\doi{10.21105/joss.03900}.
\url{https://joss.theoj.org/papers/10.21105/joss.03900}.

Druskat S, Spaaks JH, Chue Hong N, Haines R, Baker J, Bliven S,
Willighagen E, Pérez-Suárez D, Konovalov A (2021).
"Citation File Format."
\doi{10.5281/zenodo.5171937}.
\url{https://github.com/citation-file-format/citation-file-format}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/cffr/}
  \item \url{https://github.com/ropensci/cffr}
  \item Report bugs at \url{https://github.com/ropensci/cffr/issues}
}

}
\author{
\strong{Maintainer}: Diego HernangC3mez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego HernangC3mez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item JoC#o Martins (\href{https://orcid.org/0000-0001-7961-4280}{ORCID}) [reviewer]
  \item Scott Chamberlain (\href{https://orcid.org/0000-0003-1444-9135}{ORCID}) [reviewer]
}

}
\keyword{internal}
