% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cffr.R
\name{cffr}
\alias{cffr}
\alias{as.cffr}
\title{\code{cffr} objects}
\usage{
as.cffr(x)
}
\arguments{
\item{x}{a character string for the \link{as.cffr} default method}
}
\value{
A \code{cffr} list object.
}
\description{
A class and utility methods for holding CFF information.
}
\seealso{
Other core functions: 
\code{\link{cff_create}()},
\code{\link{cff_validate}()},
\code{\link{cff_write}()}
}
\concept{core functions}
