% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{cran_to_spdx}
\alias{cran_to_spdx}
\title{Mapping between \code{License} fields and SPDX}
\format{
A \link[base:data.frame]{data frame} with 87 rows and two
variables:
\describe{
\item{\code{LICENSE}}{A valid \code{License} string on CRAN.}
\item{\code{SPDX}}{A valid SPDX license identifier.}
}
}
\source{
\url{https://spdx.org/licenses/}.
}
\description{
A dataset containing the mapping between the \code{License} strings observed
in CRAN packages and their approximate match in the
\href{https://spdx.org/licenses/}{SPDX License List}.
}
\examples{
data("cran_to_spdx")

head(cran_to_spdx, 20)
}
\seealso{
\emph{Writing R Extensions}, \href{https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Licensing}{Licensing section}.
}
\keyword{datasets}
