% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cff.R
\name{create_cff}
\alias{create_cff}
\title{THIS DESC NEEDS TO BE FIXED}
\usage{
create_cff(path = ".")
}
\arguments{
\item{path}{Path to a \code{DESCRIPTION} file}
}
\value{
A character value ready for writing into \code{yaml} format.
}
\description{
Extract information from a \code{DESCRIPTION} file and creates an object
to be converted into a \code{CITATION.cff} file.
}
\seealso{
\code{\link[yaml:as.yaml]{yaml::as.yaml()}}

\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Guide to Citation File Format schema version 1.2.0}
}
