% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{cff_parse_citation}
\alias{cff_parse_citation}
\title{Previous API: Parse a \code{bibentry} to \code{cff}}
\usage{
cff_parse_citation(bib)
}
\arguments{
\item{bib}{A \code{bibentry} object.}
}
\value{
A \code{bibentry} in format \code{cff}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{\link[=as_cff.bibentry]{as_cff.bibentry()}} method
}
\examples{

bib <- citation("base")
bib


# To cff
bib_to_cff <- as_cff(bib)
bib_to_cff

}
\seealso{
\code{\link[=as_cff.bibentry]{as_cff.bibentry()}}

Other deprecated functions:
\code{\link{cff_extract_to_bibtex}()},
\code{\link{cff_from_bibtex}()},
\code{\link{cff_parse_person}()},
\code{\link{write_bib}()}
}
\concept{deprecated}
\keyword{internal}
