% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\encoding{UTF-8}
\name{cff_from_bibtex}
\alias{cff_from_bibtex}
\title{Previous API: create a \code{\link{cff}} object from BibTeX entries}
\usage{
cff_from_bibtex(x, encoding = "UTF-8", ...)
}
\arguments{
\item{x}{The source used to generate the
\code{\link{cff}} object. It must be a \code{character} object indicating either:
\itemize{
\item The path to a BibTeX file.
\item A character vector with complete BibTeX entries. See \strong{Examples}.
}}

\item{encoding}{Encoding to be assumed for \code{x}. See \code{\link[base:readLines]{base::readLines()}}.}

\item{...}{Arguments passed to \code{\link[=cff_read_bib]{cff_read_bib()}}.}
}
\value{
A \code{\link{cff_ref_lst}} object. See \code{\link[=cff_read_bib]{cff_read_bib()}} for reading \verb{*.bib} files and
\code{\link[=cff_read_bib_text]{cff_read_bib_text()}} for reading BibTeX entries supplied as a \code{character}
vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use either \code{\link[=cff_read_bib]{cff_read_bib()}} or
\code{\link[=cff_read_bib_text]{cff_read_bib_text()}}.
}
\examples{
if (requireNamespace("bibtex", quietly = TRUE)) {
  x <- c(
    "@book{einstein1921,
    title        = {Relativity: The Special and the General Theory},
    author       = {Einstein, Albert},
    year         = 1920,
    publisher    = {Henry Holt and Company},
    address      = {London, United Kingdom},
    isbn         = 9781587340925
}",
    "@misc{misc-full,
    title        = {Handing out random pamphlets in airports},
    author       = {Joe-Bob Missilany},
    year         = 1984,
    month        = oct,
    note         = {This is a full MISC entry},
    howpublished = {Handed out at O'Hare}
}"
  )

  cff_read_bib_text(x)

  # From a file.

  x2 <- system.file("examples/example.bib", package = "cffr")
  cff_read_bib(x2)
}
}
\seealso{
Deprecated functions:
\code{\link[=cff_extract_to_bibtex]{cff_extract_to_bibtex()}},
\code{\link[=cff_parse_citation]{cff_parse_citation()}},
\code{\link[=cff_parse_person]{cff_parse_person()}},
\code{\link[=write_bib]{write_bib()}}
}
\concept{deprecated}
\keyword{internal}
