% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\encoding{UTF-8}
\name{write_bib}
\alias{write_bib}
\alias{write_citation}
\title{Previous API: write files}
\usage{
write_bib(x, file = tempfile(), append = FALSE, verbose = TRUE, ascii = FALSE)

write_citation(
  x,
  file = "./inst/CITATION",
  append = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{\link[utils:bibentry]{bibentry}} or a \code{\link{cff}} object.}

\item{file}{Name of the file to be created. If \code{NULL}, the lines are
displayed instead.}

\item{append}{A logical value. If \code{TRUE}, append entries to an existing file.}

\item{verbose}{A logical value. If \code{TRUE}, display informative messages.}

\item{ascii}{A logical value. If \code{TRUE}, write entries using ASCII
characters only.}
}
\value{
Invisibly returns \code{NULL}. This function is called for its side effect of
writing a file or displaying its contents.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[=cff_write_bib]{cff_write_bib()}} or
\code{\link[=cff_write_citation]{cff_write_citation()}}.
}
\examples{
bib <- bibentry("Misc",
  title = "My title",
  author = "Fran Pérez"
)

my_temp_bib <- tempfile(fileext = ".bib")

cff_write_bib(bib, file = my_temp_bib)

cat(readLines(my_temp_bib), sep = "\n")

cff_write_bib(bib, file = my_temp_bib, ascii = TRUE, append = TRUE)

cat(readLines(my_temp_bib), sep = "\n")
}
\seealso{
\itemize{
\item \code{\link[=cff_write_bib]{cff_write_bib()}} for writing \verb{*.bib} files.
\item \code{\link[=cff_write_citation]{cff_write_citation()}} for writing \R \code{CITATION} files.
}

Deprecated functions:
\code{\link[=cff_extract_to_bibtex]{cff_extract_to_bibtex()}},
\code{\link[=cff_from_bibtex]{cff_from_bibtex()}},
\code{\link[=cff_parse_citation]{cff_parse_citation()}},
\code{\link[=cff_parse_person]{cff_parse_person()}}
}
\concept{deprecated}
\keyword{internal}
