% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff-methods.R
\name{toBibtex.cff}
\alias{toBibtex.cff}
\title{Converting \code{\link[=cff-class]{cff}} objects to BibTeX}
\usage{
\method{toBibtex}{cff}(object, ..., what = c("preferred", "references", "all"))
}
\arguments{
\item{object}{\code{cff} object.}

\item{...}{Arguments passed to \code{\link[utils:toLatex]{utils::toBibtex()}}.}

\item{what}{Fields to extract. The value could be:
\itemize{
\item \code{preferred}: This would create a single entry with the main citation
info of the package.
\item \code{references}: Extract all the entries on \code{references}.
\item \code{all}: A combination of the previous two options. This would extract
both the preferred citation info and the references.
}}
}
\description{
This method convert \code{\link[=cff-class]{cff}} objects to \code{Bibtex} objects as
provided by \code{\link[utils:toLatex]{utils::toBibtex()}}. These objects are character vectors with
BibTeX markup.
}
\examples{
a_cff <- cff_create("cffr")
a_cff$`preferred-citation`

toBibtex(a_cff, "preferred")
}
\seealso{
\code{\link[utils:toLatex]{utils::toBibtex()}}

Other functions for working with BibTeX format:
\code{\link{cff_read}()},
\code{\link{cff_read_biblines}()},
\code{\link{cff_to_bibentry}()},
\code{\link{encoded_utf_to_latex}()},
\code{\link{write_bib}()},
\code{\link{write_citation}()}
}
\concept{bibtex}
\concept{s3method}
\keyword{internal}
