% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_bibentry.R, R/cff-methods.R
\name{as_bibentry}
\alias{as_bibentry}
\alias{toBibtex.cff}
\title{Create \code{bibentry} objects from several sources}
\usage{
as_bibentry(x, what = c("preferred", "references", "all"))

\method{toBibtex}{cff}(object, ..., what = c("preferred", "references", "all"))
}
\arguments{
\item{x}{The source that would be used for generating
the \code{\link[=bibentry]{bibentry()}} object via \code{cff}. It could be:
\itemize{
\item A missing value. That would retrieve the DESCRIPTION
file on your in-development package.
\item An existing \code{\link{cff}} object,
\item Path to a CITATION.cff file (\code{"*/CITATION.cff*"}),
\item The name of an installed package (\code{"jsonlite"}), or
\item Path to a DESCRIPTION file (\code{"*/DESCRIPTION*"}).
}}

\item{what}{Fields to extract. The value could be:
\itemize{
\item \code{preferred}: This would create a single entry with the main citation
info of the package.
\item \code{references}: Extract all the entries on \code{references}.
\item \code{all}: A combination of the previous two options. This would extract
both the preferred citation info and the references.
}}

\item{object}{\code{cff} object.}

\item{...}{Arguments passed to \code{\link[utils:toLatex]{utils::toBibtex()}}.}
}
\value{
\code{as_bibentry()} returns s \code{bibentry} object (or a list of \code{bibentry}
objects).

\code{toBibtex.cff()} returns a \code{Bibtex} object that can be printed as BibTeX
markup.
}
\description{
This function creates \code{bibentry} objects (see \code{\link[utils:bibentry]{utils::bibentry()}}) from
different metadata sources (\code{cff} objects, \code{DESCRIPTION} files, etc.). Note
that a \strong{R} \code{bibentry} object is the representation of a BibTeX entry,
see \strong{Examples}

The function tries to map the information of the source \code{x} into a \code{cff}
object and performs a mapping of the metadata to BibTeX, according to
\code{vignette("bibtex_cff", "cffr")}.

Additionally, it is also provided a method for \code{\link[=toBibtex]{toBibtex()}}, that can
convert \code{\link[=cff-class]{cff}} objects to \code{Bibtex} objects as
provided by \code{\link[utils:toLatex]{utils::toBibtex()}}. These objects are character vectors with
BibTeX markup.
}
\examples{
\donttest{
# From a cff object
cff_object <- cff()

cff_object

# bibentry object
bib <- as_bibentry(cff_object)

class(bib)

bib

# Print as bibtex

toBibtex(bib)

# Thanks to the S3 Method we can also do

toBibtex(cff_object)

# From a CITATION.cff file with options

path <- system.file("examples/CITATION_complete.cff", package = "cffr")
cff_file <- as_bibentry(path, what = "all")

toBibtex(cff_file)

# For an installed package

installed_package <- as_bibentry("jsonvalidate")

toBibtex(installed_package)


# Use a DESCRIPTION file

path2 <- system.file("examples/DESCRIPTION_gitlab", package = "cffr")
desc_file <- as_bibentry(path2)

toBibtex(desc_file)
}
}
\references{
\itemize{
\item Patashnik, Oren. "BIBTEXTING" February 1988.
\url{https://osl.ugr.es/CTAN/biblio/bibtex/base/btxdoc.pdf}.
\item Haines, R., & The Ruby Citation File Format Developers. (2021).
\emph{Ruby CFF Library (Version 0.9.0)} (Computer software).
\doi{10.5281/zenodo.1184077}.
\item Hernangomez D (2022). "BibTeX and CFF, a potential crosswalk."
\emph{The cffr package, Vignettes}. \doi{10.21105/joss.03900},
\url{https://docs.ropensci.org/cffr/articles/bibtex_cff.html}.
}
}
\seealso{
\code{\link[utils:bibentry]{utils::bibentry()}}

\code{\link[utils:toLatex]{utils::toBibtex()}}

Other functions for working with BibTeX format:
\code{\link{cff_create_bib_text}()},
\code{\link{cff_read}()},
\code{\link{cff_write_bib}()},
\code{\link{encoded_utf_to_latex}()}

Other functions for converting between \strong{R} classes:
\code{\link{cff_create_person}()}

Other S3 Methods for \code{cff}:
\code{\link{cff_create_person}()}
}
\concept{bibtex}
\concept{coercing}
\concept{s3method}
