% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cff.R
\name{as_cff}
\alias{as_cff}
\alias{as_cff.default}
\alias{as_cff.list}
\alias{as_cff.person}
\alias{as_cff.bibentry}
\title{Coerce lists, \code{person} and \code{bibentry} objects to \code{\link[=cff-class]{cff}}}
\usage{
as_cff(x, ...)

\method{as_cff}{default}(x, ...)

\method{as_cff}{list}(x, ...)

\method{as_cff}{person}(x, ...)

\method{as_cff}{bibentry}(x, ...)
}
\arguments{
\item{x}{A \code{person}, \code{bibentry} or other object that could be coerced to a
list.}

\item{...}{Additional arguments to be passed on to other methods.}
}
\description{
\code{as_cff()} turns an existing list-like \strong{R} object into a so-called
\code{\link[=cff-class]{cff}}, a list with class \code{cff}.

\code{as_cff} is an S3 generic, with methods for:
\itemize{
\item \code{person} objects as produced by \code{utils::person()}.
\item \code{bibentry} objects as produced by \code{utils::bibentry()}.
\item Default: Other inputs are first coerced with \code{\link[=as.list]{as.list()}}.
}
}
