% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-cff-person.R
\encoding{UTF-8}
\name{as_cff_person}
\alias{as_cff_person}
\alias{as_cff_person.default}
\alias{as_cff_person.person}
\alias{as_cff_person.character}
\title{Coerce \R objects to \code{\link{cff_pers_lst}} objects}
\usage{
as_cff_person(x, ...)

\method{as_cff_person}{default}(x, ...)

\method{as_cff_person}{person}(x, ...)

\method{as_cff_person}{character}(x, ...)
}
\arguments{
\item{x}{Any \R object.}

\item{...}{Ignored by this method.}
}
\value{
\code{as_cff_person()} returns an object of classes
\code{\link[=cff_pers_lst]{cff_pers_lst, cff}} as defined by \code{definitions.person} or
\code{definitions.entity} specified in the following guide:
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema guide}.
Each element of the \code{cff_pers_lst} object has classes
\code{\link[=cff_pers]{cff_pers, cff}}.
}
\description{
\code{as_cff_person()} turns an existing list-like \R object into a
\code{\link{cff_pers_lst}} object representing a list of \code{definitions.person} or
\code{definitions.entity}, as defined in the following guide:
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema guide}.

\code{as_cff_person()} is an S3 generic, with methods for:
\itemize{
\item \code{person}: Objects created with \code{\link[utils:person]{utils::person()}}.
\item \code{character}: Strings with the definition for one or more authors, using
the standard BibTeX notation (see Markey, 2009) and related formats, such
as the output of \code{\link[base:format]{base::format()}} for \code{person} objects (see
\code{\link[utils:person]{format.person()}}).
\item Default: Other inputs are first coerced with \code{\link[base:as.character]{base::as.character()}}.
}

The inverse transformation (\code{cff_pers_lst} to \code{person}) can be performed
with the methods \code{\link[=as.person.cff_pers]{as.person.cff_pers()}} and \code{\link[=as.person.cff_pers_lst]{as.person.cff_pers_lst()}}.
}
\details{
\code{as_cff_person()} recognizes whether the input should be converted with the
CFF reference for \code{definitions.person} or \code{definitions.entity}.

\code{as_cff_person()} uses a custom algorithm that parses names as explained in
Section 11 of "Tame the BeaST" (Markey, 2009) (see also Decoret, 2007):
\itemize{
\item \verb{First von Last}.
\item \verb{von Last, First}.
\item \verb{von Last, Jr, First}.
}

Mapping is performed as follows:
\itemize{
\item \code{First} is mapped to the CFF key \code{given-names}.
\item \code{von} is mapped to the CFF key \code{name-particle}.
\item \code{Last} is mapped to the CFF key \code{family-names}.
\item \code{Jr} is mapped to the CFF key \code{name-suffix}.
}

For entities, the entire \code{character} is mapped to \code{name}.
We recommend "protecting" entity names with \code{{}}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Avoid unprotected entity names.
entity <- "Elephant and Castle"
as_cff_person(entity)
- name: Elephant
- name: Castle

# Protect entity names with braces.
entity_protect <- "\{Elephant and Castle\}"
as_cff_person(entity_protect)
- name: Elephant and Castle
}\if{html}{\out{</div>}}

\code{as_cff_person()} attempts to extract as much information as possible.
For \code{character} strings from \code{\link[utils:person]{format.person()}}, the
email and ORCID are also extracted.
}
\examples{
# Create a person object.
a_person <- person(
  given = "First", family = "Author",
  role = c("aut", "cre"),
  email = "first.last@example.com", comment = c(
    ORCID = "0000-0001-8457-4658",
    affiliation = "An affiliation"
  )
)

a_person

cff_person <- as_cff_person(a_person)

# Classes `cff_pers_lst` and `cff`.
class(cff_person)

# Each element has classes `cff_pers` and `cff`.
class(cff_person[[1]])

# Print.
cff_person

# Back to person object with S3 method.
as.person(cff_person)

# Coerce a string.
a_str <- paste0(
  "Julio Iglesias <fake@email.com> ",
  "(city: Miami, region: California, country: US)"
)
as_cff_person(a_str)

# Several persons.
persons <- c(
  person("Clark", "Kent", comment = c(affiliation = "Daily Planet")),
  person("Lois", "Lane"), person("Oscorp Inc.")
)

a_cff <- as_cff_person(persons)

a_cff

# Print as BibTeX with the method.
toBibtex(a_cff)

# Or as a `person` object.
as.person(a_cff)

# Or use BibTeX style as input.

x <- "Frank Sinatra and Dean Martin and Davis, Jr., Sammy and Joey Bishop"

as_cff_person(x)

as_cff_person("Herbert von Karajan")

toBibtex(as_cff_person("Herbert von Karajan"))
}
\references{
Patashnik O (1988). "BIBTEXing."
\url{https://osl.ugr.es/CTAN/biblio/bibtex/base/btxdoc.pdf}.

Markey N (2009). \emph{Tame the BeaST: The B to X of BibTeX}.
\url{https://osl.ugr.es/CTAN/info/bibtex/tamethebeast/ttb_en.pdf}.

Decoret X (2007). "A summary of BibTeX." \url{https://maverick.inria.fr/~Xavier.Decoret/resources/xdkbibtex/bibtex_summary.html#names}.
}
\seealso{
Examples in \code{vignette("cffr", package = "cffr")} and \code{\link[utils:person]{utils::person()}}.
Learn more about the \code{cff_pers_lst} and \code{cff_pers} classes in \link{cff_class}.

Convert between \R classes:
\code{\link[=as_bibentry]{as_bibentry()}},
\code{\link[=as_cff]{as_cff()}},
\code{\link{cff_class}}
}
\concept{conversions}
