% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{cff-class}
\alias{cff-class}
\title{The \code{cff} class}
\description{
\subsection{The \code{cff} class}{

\strong{cffr} implements a S3 object with \code{\link[=class]{class()}} \code{cff}, that it is used to
represent the information of a \verb{*.cff} file in \strong{R}.

Under the hood, a \code{cff} object is simply a named \code{\link{list}} to which we added
additional methods, most notably \code{\link[=print]{print()}} and \code{\link[=as_cff]{as_cff()}}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{a_named_list <- list(
  first = "I", second = "am", third = "a", fourth = "list",
  fifth = "with", sixth = "names", "none" = NULL
)


dput(a_named_list)
#> list(first = "I", second = "am", third = "a", fourth = "list", 
#>     fifth = "with", sixth = "names", none = NULL)

# Default print
a_named_list
#> $first
#> [1] "I"
#> 
#> $second
#> [1] "am"
#> 
#> $third
#> [1] "a"
#> 
#> $fourth
#> [1] "list"
#> 
#> $fifth
#> [1] "with"
#> 
#> $sixth
#> [1] "names"
#> 
#> $none
#> NULL

# But
a_cff_object <- as_cff(a_named_list)

class(a_cff_object)
#> [1] "cff"  "list"

a_cff_object
#> first: I
#> second: am
#> third: a
#> fourth: list
#> fifth: with
#> sixth: names

dput(a_cff_object)
#> structure(list(first = "I", second = "am", third = "a", fourth = "list", 
#>     fifth = "with", sixth = "names"), class = c("cff", "list"
#> ))
}\if{html}{\out{</div>}}

\code{\link[=as_cff]{as_cff()}} not only converts a \code{list} to \code{cff} but also removes items (known
as \code{keys} in CFF terminology) that are \code{NULL} or \code{NA}.
}

\subsection{Valid \code{cff} objects}{

Although \code{a_cff_object} is a \code{cff} object, it is not valid, since it does not
comply with the validation rules of the CFF specification.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{cff_validate(a_cff_object)
#> == Validating cff ==============================================================
#> x Oops! This <cff> has the following errors:
#> * cff.authors: is required
#> * cff["cff-version"]: is required
#> * cff.message: is required
#> * cff.title: is required
}\if{html}{\out{</div>}}

\code{\link[=cff_validate]{cff_validate()}} gives minimal messages of what's wrong with our \code{cff} and
(invisibly) returns the result of the validation (\code{TRUE/FALSE}).
}

\subsection{Other methods}{
\subsection{\code{\link[=as.data.frame]{as.data.frame()}}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{minimal_cff <- cff()

minimal_cff
#> authors:
#> - family-names: Doe
#>   given-names: John
#> cff-version: 1.2.0
#> message: If you use this software, please cite it using these metadata.
#> title: My Research Software

as_df <- as.data.frame(minimal_cff)

class(as_df)
#> [1] "data.frame"

t(as_df)
#>                         [,1]                                                            
#> authors.00.family_names "Doe"                                                           
#> authors.00.given_names  "John"                                                          
#> cff_version             "1.2.0"                                                         
#> message                 "If you use this software, please cite it using these metadata."
#> title                   "My Research Software"
}\if{html}{\out{</div>}}
}

}

\subsection{\code{\link[=c]{c()}}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{new_keys <- c("date-released" = "2020-01-31", abstract = "Minimal example")

c(minimal_cff, new_keys)
#> authors:
#> - family-names: Doe
#>   given-names: John
#> cff-version: 1.2.0
#> message: If you use this software, please cite it using these metadata.
#> title: My Research Software
#> date-released: '2020-01-31'
#> abstract: Minimal example
}\if{html}{\out{</div>}}
}

\subsection{\code{\link[=as.person]{as.person()}}}{

Special case for those CFF keys that are person-like.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{path <- system.file("examples/CITATION_complete.cff", package = "cffr")
the_cff <- cff_read(path)

the_cff$authors
#> - family-names: Real Person
#>   given-names: One Truly
#>   name-particle: van der
#>   name-suffix: IV
#>   alias: Citey
#>   affiliation: Excellent University, Niceplace, Arcadia
#>   address: 22 Acacia Avenue
#>   city: Citationburgh
#>   region: Renfrewshire
#>   post-code: C13 7X7
#>   country: GB
#>   orcid: https://orcid.org/0000-0001-2345-6789
#>   email: project@entity.com
#>   tel: +44(0)141-323 4567
#>   fax: +44(0)141-323 45678
#>   website: https://www.entity-project-team.io
#> - name: Entity Project Team Conference entity
#>   address: 22 Acacia Avenue
#>   city: Citationburgh
#>   region: Renfrewshire
#>   post-code: C13 7X7
#>   country: GB
#>   orcid: https://orcid.org/0000-0001-2345-6789
#>   email: project@entity.com
#>   tel: +44(0)141-323 4567
#>   fax: +44(0)141-323 45678
#>   website: https://www.entity-project-team.io
#>   date-start: '2017-01-01'
#>   date-end: '2017-01-31'
#>   location: The team garage

as.person(the_cff$authors)
#> [1] "One Truly van der Real Person IV <project@entity.com> (Citey, Excellent University, Niceplace, Arcadia, 22 Acacia Avenue, Citationburgh, Renfrewshire, C13 7X7, GB, <https://orcid.org/0000-0001-2345-6789>, +44(0)141-323 4567, +44(0)141-323 45678, https://www.entity-project-team.io)"
#> [2] "Entity Project Team Conference entity <project@entity.com> (22 Acacia Avenue, Citationburgh, Renfrewshire, C13 7X7, GB, <https://orcid.org/0000-0001-2345-6789>, +44(0)141-323 4567, +44(0)141-323 45678, https://www.entity-project-team.io, 2017-01-01, 2017-01-31, The team garage)"
}\if{html}{\out{</div>}}
}
}
\seealso{
\code{vignette("cffr", "cffr")} for a hands-on example.

Other S3 Methods for \code{cff}:
\code{\link{as_bibentry}()},
\code{\link{as_cff}()},
\code{\link{as_cff_person}()}
}
\concept{s3method}
\keyword{internal}
