% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff.R
\name{cff}
\alias{cff}
\alias{cff_modify}
\title{Read and manipulate \code{cff} objects}
\usage{
cff(path, ...)
}
\arguments{
\item{path}{The path to a \code{CITATION.cff} file.}

\item{...}{Named arguments to be used for creating a \code{cff} object. See
\strong{Details}.}
}
\value{
A \code{cff} object. Under the hood, a \code{cff} object is a regular \code{list} object
with a special \code{\link[=print]{print()}} method.
}
\description{
A class and utility methods for reading, creating and holding CFF
information. See \code{\linkS4class{cff}} to learn more about \code{cff} objects.
}
\details{
This object can be manipulated using \code{\link[=cff_create]{cff_create()}}.

\strong{Note that} this function reads \code{CITATION.cff} files. If you want to
create \code{cff} objects from DESCRIPTION files use \code{\link[=cff_create]{cff_create()}}.

If no additional \code{...} parameters are supplied (the default behavior),
a minimal valid \code{cff} object is created. Valid parameters are those
specified on \code{\link[=cff_schema_keys]{cff_schema_keys()}}:\tabular{l}{
   \strong{valid cff keys} \cr
   cff-version \cr
   message \cr
   type \cr
   license \cr
   title \cr
   version \cr
   doi \cr
   abstract \cr
   authors \cr
   preferred-citation \cr
   repository \cr
   repository-artifact \cr
   repository-code \cr
   url \cr
   date-released \cr
   contact \cr
   keywords \cr
   references \cr
   commit \cr
   identifiers \cr
   license-url \cr
}
}
\examples{

# Blank cff
cff()

# From file
cff_read(system.file("examples/CITATION_basic.cff",
  package = "cffr"
))

# Use custom params
test <- cff(
  title = "Manipulating files",
  keywords = c("A", "new", "list", "of", "keywords"),
  authors = as_cff_person("New author")
)
test
\donttest{
# Would fail
cff_validate(test)


# Modify with cff_create
new <- cff_create(test, keys = list(
  "cff-version" = "1.2.0",
  message = "A blank file"
))
new

# Would pass
cff_validate(new)
}
}
\seealso{
Other core functions of \CRANpkg{cffr}:
\code{\link{cff_create}()},
\code{\link{cff_validate}()}
}
\concept{core}
