% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff.R
\name{cff}
\alias{cff}
\alias{as.cff}
\title{\code{cff} objects}
\usage{
as.cff(x)
}
\arguments{
\item{x}{a character string for the \code{\link{as.cff}} default method}
}
\value{
A \code{cff} list object.
}
\description{
A class and utility methods for holding CFF information.
}
\seealso{
Other core functions: 
\code{\link{cff_create}()},
\code{\link{cff_validate}()},
\code{\link{cff_write}()}
}
\concept{core functions}
