% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff.R
\name{cff}
\alias{cff}
\alias{as.cff}
\title{\code{cff} objects}
\usage{
cff(path, ...)

as.cff(x)
}
\arguments{
\item{path}{The path to a \code{CITATION.cff} file.}

\item{...}{Named arguments to be used for creating a \code{\link{cff}} object.}

\item{x}{a character string for the \code{\link{as.cff}} default method}
}
\value{
A \code{cff} list object.
}
\description{
A class and utility methods for reading, creating and holding CFF
information.
}
\details{
If no parameters are supplied (the default behavior), a minimal
valid \code{cff} object is created.

This object can be manipulated using \code{\link[=cff_create]{cff_create()}}.

\strong{Note that} this function reads \code{CITATION.cff} files. If you want to
create \code{cff} objects from DESCRIPTION files use \code{\link[=cff_create]{cff_create()}}
}
\examples{

# Blank cff
cff()

# From file
cff(system.file("examples/CITATION_basic.cff",
  package = "cffr"
))

# Use custom params
test <- cff(
  message = "A blank file",
  keywords = c("A", "new", "list", "of", "keywords"),
  authors = list(cff_parse_person("New author"))
)
test

# Would fail
cff_validate(test)


# Modify with cff_create
new <- cff_create(test, keys = list(
  "cff-version" = "1.2.0",
  title = "Manipulating files"
))
new

# Would pass
cff_validate(new)

as.cff(list(
  "cff-version" = "1.2.0",
  title = "Manipulating files"
))
}
\seealso{
Other core functions: 
\code{\link{cff_create}()},
\code{\link{cff_validate}()},
\code{\link{cff_write}()}
}
\concept{core functions}
