% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R, R/methods.R
\encoding{UTF-8}
\name{cff_class}
\alias{cff_class}
\alias{cff_ref}
\alias{cff_ref_lst}
\alias{cff_pers}
\alias{cff_pers_lst}
\alias{print.cff}
\alias{c.cff}
\alias{[.cff_ref_lst}
\alias{[.cff_pers_lst}
\alias{as.data.frame.cff}
\alias{as.data.frame.cff_pers_lst}
\alias{as.data.frame.cff_pers}
\alias{as.data.frame.cff_ref_lst}
\alias{as.data.frame.cff_ref}
\alias{head.cff}
\alias{tail.cff}
\alias{as.list.cff}
\alias{toBibtex.cff}
\alias{toBibtex.cff_ref_lst}
\alias{toBibtex.cff_ref}
\alias{toBibtex.cff_pers_lst}
\alias{toBibtex.cff_pers}
\alias{as.person.cff_pers}
\alias{as.person.cff_pers_lst}
\alias{as.person.cff}
\alias{as.person.cff_ref}
\alias{as.person.cff_ref_lst}
\title{The \code{cff} class}
\value{
A documentation topic describing the \code{cff} class and subclasses.
}
\description{
\CRANpkg{cffr} implements an S3 object with \code{\link[base:class]{base::class()}} \code{cff}, which
represents the information in a \code{CITATION.cff} file in \R.

Under the hood, a \code{\link{cff}} object is a named \code{\link{list}} with additional methods,
most notably \code{\link[=print]{print()}} and \code{\link[=as_cff]{as_cff()}}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{a_named_list <- list(
  first = "I", second = "am", third = "a", fourth = "list",
  fifth = "with", sixth = "names", "none" = NULL
)

# Convert to `cff`.
a_cff_object <- as_cff(a_named_list)

class(a_cff_object)
#> [1] "cff"

a_cff_object
#> first: I
#> second: am
#> third: a
#> fourth: list
#> fifth: with
#> sixth: names

is.list(a_cff_object)
#> [1] TRUE
}\if{html}{\out{</div>}}

\code{\link[=as_cff]{as_cff()}} converts a \code{\link{list}} to \code{\link{cff}} and removes elements that are \code{NULL}
or \code{NA}. Named elements are known as keys in CFF terminology.
\subsection{Subclasses}{

\CRANpkg{cffr} implements two special subclasses of \code{\link{cff}}, which represent
two types of objects defined in the \href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema}:
\itemize{
\item \code{definitions.person} and \code{definitions.entity}: CFF definitions for sublists
representing persons or entities. In \CRANpkg{cffr}, the subclass
\code{\link{cff_pers_lst}} collects an array of \code{definitions.person} or
\code{definitions.entity} objects, where each individual person or entity has
subclass \code{\link{cff_pers}}.
\item \code{definitions.reference}: The CFF definition for references to related works
and other software used in the main work described by the \code{CITATION.cff} file.
In \CRANpkg{cffr}, the subclass \code{\link{cff_ref_lst}} represents arrays of
\code{definitions.reference} objects, where each element has subclass
\code{\link{cff_ref}}.
}

These subclasses are not complete, valid \code{\link{cff}} objects. They adapt
information to the structures required by the CFF schema.

The examples below use \code{\link[utils:person]{utils::person()}}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{## Array.
two_persons <- as_cff_person(
  c(
    person("A", "person", comment = c(ORCID = "0000-0001-8457-4658")),
    person("An entity", email = "fake@gmail.com")
  )
)

two_persons
#> - family-names: person
#>   given-names: A
#>   orcid: https://orcid.org/0000-0001-8457-4658
#> - name: An entity
#>   email: fake@gmail.com

class(two_persons)
#> [1] "cff_pers_lst" "cff"

# Single element.

two_persons[[1]]
#> family-names: person
#> given-names: A
#> orcid: https://orcid.org/0000-0001-8457-4658

class(two_persons[[1]])
#> [1] "cff_pers" "cff"

# Array of references.

cit <- c(citation(), citation("jsonlite"))

ref_list <- as_cff(cit)

ref_list
#> - type: manual
#>   title: 'R: A Language and Environment for Statistical Computing'
#>   authors:
#>   - name: R Core Team
#>     website: https://ror.org/02zz1nj61
#>   institution:
#>     name: R Foundation for Statistical Computing
#>     website: https://ror.org/05qewa988
#>     address: Vienna, Austria
#>   year: '2026'
#>   doi: 10.32614/R.manuals
#>   url: https://www.R-project.org/
#> - type: article
#>   title: 'The jsonlite Package: A Practical and Consistent Mapping Between JSON Data
#>     and R Objects'
#>   authors:
#>   - family-names: Ooms
#>     given-names: Jeroen
#>   journal: arXiv:1403.2805 [stat.CO]
#>   year: '2014'
#>   url: https://arxiv.org/abs/1403.2805

class(ref_list)
#> [1] "cff_ref_lst" "cff"

# Single element.

ref_list[[1]]
#> type: manual
#> title: 'R: A Language and Environment for Statistical Computing'
#> authors:
#> - name: R Core Team
#>   website: https://ror.org/02zz1nj61
#> institution:
#>   name: R Foundation for Statistical Computing
#>   website: https://ror.org/05qewa988
#>   address: Vienna, Austria
#> year: '2026'
#> doi: 10.32614/R.manuals
#> url: https://www.R-project.org/

class(ref_list[[1]])
#> [1] "cff_ref" "cff"
}\if{html}{\out{</div>}}
}

\subsection{Valid \code{\link{cff}} objects}{

Creating a \code{\link{cff}} object does not ensure its validity against the \href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{class(a_cff_object)
#> [1] "cff"

cff_validate(a_cff_object)
#> == Validating CFF ==============================================================
#> x Validation failed. This <cff> object has the following errors:
#> * cff/: must have required property 'authors'
#> * cff/: must have required property 'cff-version'
#> * cff/: must have required property 'message'
#> * cff/: must have required property 'title'
#> * cff/: must NOT have additional properties
#> * cff/: must NOT have additional properties
#> * cff/: must NOT have additional properties
#> * cff/: must NOT have additional properties
#> * cff/: must NOT have additional properties
#> * cff/: must NOT have additional properties
}\if{html}{\out{</div>}}

\code{\link[=cff_validate]{cff_validate()}} reports validation problems and invisibly returns \code{TRUE} or
\code{FALSE}.

We can use \code{\link[=cff_modify]{cff_modify()}} to add more keys:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{cff_valid <- cff_modify(a_cff_object,
  authors = as_cff_person("\{James and James\}"),
  cff_version = "1.2.0",
  message = "Hi there",
  title = "My title"
)

# Remove invalid keys.
cff_valid <- as_cff(cff_valid[names(cff_valid) \%in\% cff_schema_keys()])

cff_valid
#> authors:
#> - name: James and James
#> cff-version: 1.2.0
#> message: Hi there
#> title: My title

cff_validate(cff_valid)
#> == Validating CFF ==============================================================
#> v This <cff> object is valid.
}\if{html}{\out{</div>}}
}

\subsection{Base methods provided by \CRANpkg{cffr}}{

\CRANpkg{cffr} provides additional S3 methods for common coercion functions from
the base and utils namespaces.
\subsection{\code{\link[=as.data.frame]{as.data.frame()}}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{minimal_cff <- cff()

minimal_cff
#> cff-version: 1.2.0
#> message: If you use this software, please cite it using these metadata.
#> title: My Research Software
#> authors:
#> - family-names: Doe
#>   given-names: John

as_df <- as.data.frame(minimal_cff)

class(as_df)
#> [1] "data.frame"

t(as_df)
#>                         [,1]                                                            
#> cff_version             "1.2.0"                                                         
#> message                 "If you use this software, please cite it using these metadata."
#> title                   "My Research Software"                                          
#> authors.00.family_names "Doe"                                                           
#> authors.00.given_names  "John"
}\if{html}{\out{</div>}}
}

\subsection{\code{\link[=c]{c()}}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{new_keys <- c("date-released" = "2020-01-31", abstract = "Minimal example")

c(minimal_cff, new_keys)
#> cff-version: 1.2.0
#> message: If you use this software, please cite it using these metadata.
#> title: My Research Software
#> authors:
#> - family-names: Doe
#>   given-names: John
#> date-released: '2020-01-31'
#> abstract: Minimal example
}\if{html}{\out{</div>}}
}

\subsection{\code{\link[=as.list]{as.list()}}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{as.list(minimal_cff)
#> $`cff-version`
#> [1] "1.2.0"
#> 
#> $message
#> [1] "If you use this software, please cite it using these metadata."
#> 
#> $title
#> [1] "My Research Software"
#> 
#> $authors
#> $authors[[1]]
#> $authors[[1]]$`family-names`
#> [1] "Doe"
#> 
#> $authors[[1]]$`given-names`
#> [1] "John"
}\if{html}{\out{</div>}}
}

\subsection{\code{\link[=as.person]{as.person()}}}{

Only for \code{\link{cff_pers_lst}} and \code{\link{cff_pers}} objects:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{as.person(two_persons)
#> [1] "A person (ORCID: <https://orcid.org/0000-0001-8457-4658>)"
#> [2] "An entity <fake@gmail.com>"
}\if{html}{\out{</div>}}
}

\subsection{\code{\link[=toBibtex]{toBibtex()}}}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# For `cff`.
toBibtex(minimal_cff)
#> @Misc\{doe,
#>   title = \{My Research Software\},
#>   author = \{John Doe\},
#> \}

# For `cff_ref` and `cff_ref_lst`.
toBibtex(cit)
#> @Manual\{,
#>   title = \{R: A Language and Environment for Statistical Computing\},
#>   author = \{\{R Core Team\}\},
#>   organization = \{R Foundation for Statistical Computing\},
#>   address = \{Vienna, Austria\},
#>   year = \{2026\},
#>   doi = \{10.32614/R.manuals\},
#>   url = \{https://www.R-project.org/\},
#> \}
#> 
#> @Article\{,
#>   title = \{The jsonlite Package: A Practical and Consistent Mapping Between JSON Data and R Objects\},
#>   author = \{Jeroen Ooms\},
#>   journal = \{arXiv:1403.2805 [stat.CO]\},
#>   year = \{2014\},
#>   url = \{https://arxiv.org/abs/1403.2805\},
#> \}

# For `cff_pers` and `cff_pers_lst`.
toBibtex(two_persons)
#> person, A and \{An entity\}
}\if{html}{\out{</div>}}
}

}
}
\references{
Wickham H (2019). "S3." In \emph{Advanced R}, 2nd edition.
Chapman and Hall/CRC.
\doi{10.1201/9781351201315}.
\url{https://adv-r.hadley.nz/s3.html}.
}
\seealso{
Convert between \R classes:
\code{\link[=as_bibentry]{as_bibentry()}},
\code{\link[=as_cff]{as_cff()}},
\code{\link[=as_cff_person]{as_cff_person()}}
}
\concept{conversions}
\keyword{internal}
