% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_create.R
\name{cff_create}
\alias{cff_create}
\title{Create a \code{ccfr} object}
\usage{
cff_create(path = ".", keys = NULL)
}
\arguments{
\item{path}{path to package root (\code{"."}), or package name (\code{"jsonlite"}), or
description file (\code{"*/DESCRIPTION*"}).}

\item{keys}{List of additional keys to add to the \code{ccfr} object. See
\strong{Details}}
}
\value{
A \code{ccfr} list object.
}
\description{
Create a \code{ccfr} object for further manipulation. Similar to
\code{\link[=cff_write]{cff_write()}}, but returns a list object rather than writing directly to a
file. See examples.
}
\details{
It is possible to add additional keys not detected by \link{cff_create} using the
\code{keys} argument. A list of valid keys can be retrieved with
\code{\link[=cff_schema_keys]{cff_schema_keys()}}.

Please refer to \href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Guide to Citation File Format schema version 1.2.0}
for additional details.
}
\examples{



# Installed package
cff_create("jsonlite")

# Demo file
demo_file <- system.file("examples/DESCRIPTION_demo", package = "cffr")
cff_create(demo_file)

# Add additional keys

newkeys <- list(
  message = "This overwrites fields",
  abstract = "New abstract",
  keywords = c("A", "new", "list", "of", "keywords"),
  authors = list(list(
    "family-names" = "Doe",
    "given-names" = "John"
  )),
  "date-released" = "1900-01-01"
)

cff_create(demo_file, keys = newkeys)

# Update a field
# New author

old <- cff_create(demo_file)

old_aut <- old$authors

# It should be a list
new_aut <- append(
  old_aut,
  list(list(
    "family-names" = "new author",
    "given-names" = "I am a",
    alias = "Hello!"
  ))
)

cff_create(demo_file, keys = list("authors" = new_aut))
}
\seealso{
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Guide to Citation File Format schema version 1.2.0},
\code{\link[=cff_write]{cff_write()}}
}
