% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_create.R
\name{cff_create}
\alias{cff_create}
\title{Create a \code{ccfr} object}
\usage{
cff_create(x = ".", keys = NULL, cff_version = "1.2.0")
}
\arguments{
\item{x}{It could be
\itemize{
\item The path to package root (\code{"."}),
\item The name of an installed package (\code{"jsonlite"}), or
\item Path to a DESCRIPTION file (\code{"*/DESCRIPTION*"}).
}}

\item{keys}{List of additional keys to add to the \code{ccfr} object. See
\strong{Details}.}

\item{cff_version}{The Citation File Format schema version that the
CITATION.cff file adheres to for providing the citation metadata.}
}
\value{
A \code{ccfr} list object.
}
\description{
Create a \code{ccfr} object for further manipulation. Similar to
\code{\link[=cff_write]{cff_write()}}, but returns a list object rather than writing directly to a
file. See examples.
}
\details{
It is possible to add additional keys not detected by \link{cff_create} using the
\code{keys} argument. A list of valid keys can be retrieved with
\code{\link[=cff_schema_keys]{cff_schema_keys()}}.

Please refer to \href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Guide to Citation File Format schema version 1.2.0}
for additional details.
}
\examples{



# Installed package
cff_create("jsonlite")

# Demo file
demo_file <- system.file("examples/DESCRIPTION_demo", package = "cffr")
cff_create(demo_file)

# Add additional keys

newkeys <- list(
  message = "This overwrites fields",
  abstract = "New abstract",
  keywords = c("A", "new", "list", "of", "keywords"),
  authors = list(cff_parse_person("New author")),
  "date-released" = "1900-01-01"
)

cff_create(demo_file, keys = newkeys)

# Update a field on a list - i,e: authors, contacts, etc.
# We are adding a new contact here

old <- cff_create(demo_file)

new_contact <- append(
  old$contact,
  list(
    cff_parse_person(person(
      given = "I am",
      family = "New Contact"
    ))
  )
)


cff_create(demo_file, keys = list("contact" = new_contact))
}
\seealso{
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Guide to Citation File Format schema version 1.2.0},
\code{\link[=cff_write]{cff_write()}}

Other \strong{cffr} core functions: 
\code{\link{cff_validate}()},
\code{\link{cff_write}()}
}
\concept{\strong{cffr} core functions}
