% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_modify.R
\name{cff_modify}
\alias{cff_modify}
\title{Modify a \code{\link{cff}} object}
\usage{
cff_modify(x, ...)
}
\arguments{
\item{x}{A \code{\link{cff}} object.}

\item{...}{Named arguments to be used for modifying \code{x}. See also \code{\link[=cff]{cff()}}.}
}
\value{
A \code{\link{cff}} object.
}
\description{
Add new keys or modify existing ones on a \code{\link{cff}} object.
}
\details{
If any key provided in \code{...} is present in \code{x}, the result would have the
key provided in \code{...}.
}
\examples{
x <- cff()
x

cff_validate(x)


x_mod <- cff_modify(x,
  contact = as_cff_person("A contact"),
  message = "This overwrites fields",
  title = "New Title",
  abstract = "New abstract",
  doi = "10.21105/joss.03900"
)
cff_validate(x_mod)

}
\seealso{
This function is wrapper of \code{\link[utils:modifyList]{utils::modifyList()}}.

See \code{\link[=cff]{cff()}} for creating \code{\link{cff}} objects from scratch.

Other core functions of \CRANpkg{cffr}:
\code{\link{cff}()},
\code{\link{cff_create}()},
\code{\link{cff_validate}()}

Other core functions of \CRANpkg{cffr}:
\code{\link{cff}()},
\code{\link{cff_create}()},
\code{\link{cff_validate}()}
}
\concept{core}
