% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_parse_person.R
\name{cff_parse_person}
\alias{cff_parse_person}
\alias{cff_parse_person_bibtex}
\title{Parse a person to \code{cff}}
\usage{
cff_parse_person(person)

cff_parse_person_bibtex(person)
}
\arguments{
\item{person}{See \strong{Details}.}
}
\value{
A \code{\link{cff}} object ready to be used on \code{\link[=cff_create]{cff_create()}}.
}
\description{
Parse a person or string to a valid format for a \code{CITATION.cff} file.
}
\details{
This is a helper function designed to help on adding or
replacing the auto-generated authors of the package. See \strong{Examples}.

\code{person} could be:
\itemize{
\item For \code{cff_parse_person()}: A \code{person} object or a character coercible to
\code{person}. See \code{\link[=person]{person()}} for details.
\item For \code{cff_parse_person_bibtex()}: A string with the definition of an author
or several authors, using the standard BiBTeX notation. See Markey (2007)
for a full explanation.
}
}
\examples{
# Parse a person object

cff_parse_person(person(
  given = "First",
  family = "Author",
  role = c("aut", "cre"),
  email = "first.last@example.com",
  comment = c(
    ORCID = "0000-0001-8457-4658",
    affiliation = "An affiliation"
  )
))

# Parse a string

cff_parse_person("Julio Iglesias <fake@email.com>")

# Several persons
persons <- c(person("Clark", "Kent"), person("Lois", "Lane"))

cff_parse_person(persons)

# Using bibtex style

x <- "Frank Sinatra and Dean Martin and Davis, Jr., Sammy and Joey Bishop"

cff_parse_person_bibtex(x)

cff_parse_person_bibtex("Herbert von Karajan")
}
\references{
Patashnik O (1988). "BIBTEXTING."
\url{https://osl.ugr.es/CTAN/biblio/bibtex/base/btxdoc.pdf}.

Markey N (2007). "Tame the BeaST" \emph{The B to X of BibTEX}, version 1.4.
\url{http://tug.ctan.org/info/bibtex/tamethebeast/ttb_en.pdf}.
}
\seealso{
\code{\link[=cff_create]{cff_create()}}, \code{vignette("cffr", "cffr")}, \code{\link[utils:person]{utils::person()}}

Other parsers: 
\code{\link{cff_parse_citation}()}
}
\concept{parsers}
