% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_read.R
\name{cff_read}
\alias{cff_read}
\alias{cff_read_cff_citation}
\alias{cff_read_description}
\alias{cff_read_bib}
\title{Read an external file as a \code{\link[=cff-class]{cff}} object}
\usage{
cff_read(path, ...)

cff_read_cff_citation(path, ...)

cff_read_description(
  path,
  cff_version = "1.2.0",
  gh_keywords = TRUE,
  authors_roles = c("aut", "cre"),
  ...
)

cff_read_cff_citation(path, ...)

cff_read_bib(path, encoding = "UTF-8", ...)
}
\arguments{
\item{path}{Path to a file}

\item{...}{Arguments to be passed to other functions.}

\item{cff_version}{The Citation File Format schema version that the
\code{CITATION.cff} file adheres to for providing the citation metadata.}

\item{gh_keywords}{Logical \code{TRUE/FALSE}. If the package is hosted on
GitHub, would you like to add the repo topics as keywords?}

\item{authors_roles}{Roles to be considered as authors of the package when
generating the \code{\link[=cff-class]{cff}} object.}

\item{encoding}{Encoding to be assumed for \code{path}. See \code{\link[=readLines]{readLines()}}.}
}
\value{
A \code{\link[=cff-class]{cff}} object.
}
\description{
Read files and convert them to \code{\link[=cff-class]{cff}} objects. Files supported
are:
\itemize{
\item \code{CITATION.cff} files
\item \code{DESCRIPTION} files
\item \strong{R} citation files (usually located in \code{inst/CITATION})
\item BibTeX files (with extension \verb{*.bib}).
}

\code{\link[=cff_read]{cff_read()}} would try to guess the type of file provided in \code{path}. However
we provide a series of alias for each specific type of file:
\itemize{
\item \code{\link[=cff_read_cff_citation]{cff_read_cff_citation()}}, that uses \code{\link[yaml:read_yaml]{yaml::read_yaml()}}.
\item \code{\link[=cff_read_description]{cff_read_description()}}, using \code{\link[desc:desc]{desc::desc()}}.
\item \code{\link[=cff_read_citation]{cff_read_citation()}} uses \code{\link[utils:citation]{utils::readCitationFile()}}.
\item \code{\link[=cff_read_bib]{cff_read_bib()}}  requires \CRANpkg{bibtex} (>= 0.5.0) and uses
\code{\link[bibtex:read.bib]{bibtex::read.bib()}}.
}
}
