% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-schema.R
\encoding{UTF-8}
\name{cff_schema}
\alias{cff_schema}
\alias{cff_schema_keys}
\alias{cff_schema_keys_license}
\alias{cff_schema_definitions_person}
\alias{cff_schema_definitions_entity}
\alias{cff_schema_definitions_refs}
\title{Inspect CFF schema values}
\source{
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema guide}.
}
\usage{
cff_schema_keys(sorted = FALSE)

cff_schema_keys_license()

cff_schema_definitions_person()

cff_schema_definitions_entity()

cff_schema_definitions_refs()
}
\arguments{
\item{sorted}{A logical value. If \code{TRUE}, arrange the keys alphabetically.}
}
\value{
A character vector with the names of valid keys for Citation File Format
version 1.2.0.
}
\description{
Helper functions with valid values for different keys and definition fields.
See the following guide:
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema guide}.
\itemize{
\item \code{\link[=cff_schema_keys]{cff_schema_keys()}} provides the valid high-level keys of the Citation
File Format.
\item \code{\link[=cff_schema_keys_license]{cff_schema_keys_license()}} provides valid
\href{https://spdx.org/licenses/}{SPDX license identifier(s)} for the
\code{CITATION.cff} file.
\item \code{\link[=cff_schema_definitions_person]{cff_schema_definitions_person()}} and \code{\link[=cff_schema_definitions_entity]{cff_schema_definitions_entity()}}
return the valid fields to include when defining a
person or entity.
\item \code{\link[=cff_schema_definitions_refs]{cff_schema_definitions_refs()}} provides the valid
keys to use in the \code{preferred-citation} and \code{references} keys.
}
}
\examples{
cff_schema_keys(sorted = TRUE)
# Valid license keys.
head(cff_schema_keys_license(), 20)
cff_schema_definitions_person()
cff_schema_definitions_entity()
cff_schema_definitions_refs()
}
\concept{schemas}
