% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff-validate.R
\encoding{UTF-8}
\name{cff_validate}
\alias{cff_validate}
\title{Validate a \code{CITATION.cff} file or a \code{\link{cff}} object}
\usage{
cff_validate(x = "CITATION.cff", verbose = TRUE)
}
\arguments{
\item{x}{A full \code{cff} object created with \code{\link[=cff_create]{cff_create()}} or the path to a
\code{CITATION.cff} file to be validated. A \verb{*.cff} file is read with
\code{\link[=cff_read_cff_citation]{cff_read_cff_citation()}}.}

\item{verbose}{A logical value. If \code{TRUE}, the function displays
informative messages.}
}
\value{
Invisibly returns a logical value indicating the validation result. On
error, the result has an \code{"errors"} attribute with the error summary. If
\code{verbose} is \code{TRUE}, the function also displays the result. See \strong{Examples}
and \code{\link[base:attr]{base::attr()}}.
}
\description{
Validate a \code{CITATION.cff} file or a \code{\link{cff}} object using the corresponding \href{https://github.com/citation-file-format/citation-file-format/blob/main/schema.json}{validation schema}.
}
\examples{
\donttest{
# Full `.cff` example.
cff_validate(system.file("examples/CITATION_complete.cff", package = "cffr"))

# Validate a `cff` object.
cffr <- cff_create("jsonlite")
class(cffr)
cff_validate(cffr)
}

# `.cff` with errors.
err_f <- system.file("examples/CITATION_error.cff", package = "cffr")
# You can manipulate the errors as a data frame.
res <- try(cff_validate(err_f))

isTRUE(res)
isFALSE(res)

# Detailed results.
df_errors <- attr(res, "errors")

df_errors[, c("field", "message")]

# An R citation file is not a `.cff` file, so it throws an error.
try(cff_validate(system.file("CITATION", package = "cffr")))
}
\seealso{
\code{\link[jsonvalidate:json_validate]{jsonvalidate::json_validate()}}, which is the function that performs the
validation.

\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Citation File Format schema guide}.

Core \CRANpkg{cffr} workflow:
\code{\link[=cff]{cff()}},
\code{\link[=cff_create]{cff_create()}},
\code{\link[=cff_modify]{cff_modify()}},
\code{\link[=cff_write]{cff_write()}}
}
\concept{core}
