% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_validate.R
\name{cff_validate}
\alias{cff_validate}
\title{Validate a CITATION.cff file or a \code{cffr} object}
\usage{
cff_validate(path = "./CITATION.cff")
}
\arguments{
\item{path}{This is expected to be either a \code{cffr} object created
with \code{\link[=cff_create]{cff_create()}} or the path to a *.cff file to be validated.}
}
\value{
A message indicating the result of the validation and an invisible
value \code{TRUE/FALSE}.
}
\description{
Validate a CITATION.cff file or a \code{cffr} object created with
\code{\link[=cff_create]{cff_create()}} using the corresponding validation
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema.json}{schema.json}.
}
\examples{
# Full .cff example
cff_validate(system.file("examples/CITATION_full.cff", package = "cffr"))

# Validate a cffr object
cffr <- cff_create("jsonlite")
class(cffr)
cff_validate(cffr)

# .cff with errors
cff_validate(system.file("examples/CITATION_error.cff", package = "cffr"))
\dontrun{
# If a CITATION file (note that is not .cff) it throws an error
cff_validate(system.file("CITATION", package = "cffr"))
}
}
\seealso{
\code{\link[=cff_create]{cff_create()}}, \code{\link[=cff_write]{cff_write()}}
}
\concept{**cffr** core functions}
\concept{cff_core}
