% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_validate.R
\name{cff_validate}
\alias{cff_validate}
\title{Validate a \code{CITATION.cff} file}
\usage{
cff_validate(path = "./CITATION.cff")
}
\arguments{
\item{path}{The path to the \code{CITATION.cff} file to validate}
}
\value{
A message indicating the result of the validation.
}
\description{
Validate a \code{CITATION.cff} file using the corresponding
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema.json}{schema.json}
}
\examples{
# Full .cff example
cff_validate(system.file("examples/CITATION_full.cff", package = "cffR"))

# .cff with errors
cff_validate(system.file("examples/CITATION_error.cff", package = "cffR"))
\dontrun{
# If a CITATION file (note that is not .cff) it throws an error
cff_validate(system.file("CITATION", package = "cffR"))
}
}
