% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_write.R
\name{cff_write}
\alias{cff_write}
\title{Write a \code{CITATION.cff} file}
\usage{
cff_write(
  x = ".",
  outfile = "CITATION.cff",
  keys = NULL,
  cff_version = "1.2.0",
  validate = TRUE
)
}
\arguments{
\item{x}{The source that would be used for generating
the \code{CITATION.cff} file. It could be:
\itemize{
\item A \code{\link{cff}} object,
\item The path to package root (\code{"."}),
\item The name of an installed package (\code{"jsonlite"}), or
\item Path to a DESCRIPTION file (\code{"*/DESCRIPTION*"}).
}}

\item{outfile}{The name and path of the \code{CITATION.cff} to be created.}

\item{keys}{List of additional keys to add to the \code{\link{cff}} object. See
\code{\link[=cff_create]{cff_create()}} for details and examples.}

\item{cff_version}{The Citation File Format schema version that the
\code{CITATION.cff} file adheres to for providing the citation metadata.}

\item{validate}{Logical \code{TRUE/FALSE}. Should the new file be validated using
\code{\link[=cff_validate]{cff_validate()}}?}
}
\value{
A \code{CITATION.cff} file and an (invisible) \code{\link{cff}} object.
}
\description{
\strong{This is the core function of the package and likely to be the only one
you would need when developing a package}.

This funxtion writes out a \code{CITATION.cff} file for a given package. This function is
basically a wrapper around \code{\link[=cff_create]{cff_create()}} to both create the \code{\link{cff}} object
and write it out to a YAML-formatted file in one command.
}
\details{
When creating and writing a CITATION.cff for the first time, the function
adds "CITATION.cff" to .Rbuildignore
}
\examples{

cff_obj <- cff_write("jsonlite", outfile = tempfile(fileext = ".cff"))

cff_obj
}
\seealso{
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Guide to Citation File Format schema version 1.2.0}.

Other core functions: 
\code{\link{cff_create}()},
\code{\link{cff_validate}()},
\code{\link{cff}()}
}
\concept{core functions}
