% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff_write.R
\name{cff_write}
\alias{cff_write}
\title{Write a CITATION.cff file}
\usage{
cff_write(
  x = ".",
  outfile = "CITATION.cff",
  keys = NULL,
  cff_version = "1.2.0",
  validate = TRUE
)
}
\arguments{
\item{x}{It could be
\itemize{
\item The path to package root (\code{"."}),
\item The name of an installed package (\code{"jsonlite"}), or
\item Path to a DESCRIPTION file (\code{"*/DESCRIPTION*"}).
}}

\item{outfile}{The name and path of the CITATION.cff to be created.}

\item{keys}{List of additional keys to add to the \code{ccfr} object. See
\code{\link[=cff_create]{cff_create()}} for details and examples.}

\item{cff_version}{The Citation File Format schema version that the
CITATION.cff file adheres to for providing the citation metadata.}

\item{validate}{Logical \code{TRUE/FALSE}. Should the new file be validated using
\code{\link[=cff_validate]{cff_validate()}}?}
}
\value{
A CITATION.cff file and an (invisible) \code{cffr} object.
}
\description{
Write out a CITATION.cff file for a given package. This function is
basically a wrapper around \code{\link[=cff_create]{cff_create()}} to both create the \code{cffr} object
and write it out to a YAML-formatted file in one command.
}
\details{
When creating and writing a CITATION.cff for the first time, the function
adds "CITATION.cff" to .Rbuildignore
}
\examples{

cff_obj <- cff_write("jsonlite", outfile = tempfile(fileext = ".cff"))

cff_obj
}
\seealso{
\href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Guide to Citation File Format schema version 1.2.0}.

Other \strong{cffr} core functions: 
\code{\link{cff_create}()},
\code{\link{cff_validate}()}
}
\concept{\strong{cffr} core functions}
