% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cff-write-misc.R
\encoding{UTF-8}
\name{cff_write_bib}
\alias{cff_write_bib}
\alias{cff_write_citation}
\title{Export \R objects to multiple file types}
\usage{
cff_write_bib(
  x,
  file = tempfile(fileext = ".bib"),
  append = FALSE,
  verbose = TRUE,
  ascii = FALSE,
  ...
)

cff_write_citation(
  x,
  file = tempfile("CITATION_"),
  append = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{\link{bibentry}} or a \code{\link{cff}} object.}

\item{file}{Name of the file to be created. If \code{NULL}, the lines are
displayed instead.}

\item{append}{A logical value. If \code{TRUE}, append entries to an existing file.}

\item{verbose}{A logical value. If \code{TRUE}, display informative messages.}

\item{ascii}{A logical value. If \code{TRUE}, write entries using ASCII
characters only.}

\item{...}{
  Arguments passed on to \code{\link{as_bibentry.cff}}, \code{\link{as_bibentry.cff_ref}}, \code{\link{as_bibentry.cff_ref_lst}}
  \describe{
    \item{\code{what}}{Fields to extract from a full \code{cff} object. It can be:
\itemize{
\item \code{preferred}: Create a single entry with the main citation information
of the package (key \code{preferred-citation}).
\item \code{references}: Extract all entries of the \code{references} key.
\item \code{all}: Extract both the \code{preferred-citation} and \code{references} keys.
}

See \code{vignette("r-cff", package = "cffr")}.}
  }}
}
\value{
Invisibly returns \code{NULL}. This function is called for its side effect of
writing a file or displaying its contents.
}
\description{
Export \R objects representing citations to specific file formats:
\itemize{
\item \code{\link[=cff_write_bib]{cff_write_bib()}} creates a \code{.bib} file.
\item \code{\link[=cff_write_citation]{cff_write_citation()}} creates an \R citation file as described in
Section 1.9 of \emph{Writing R Extensions} (R Core Team 2026).
}
}
\details{
When \code{x} is a \code{cff} object, it is converted to BibTeX using
\code{\link[=toBibtex.cff]{toBibtex.cff()}}.

For security reasons, if the file already exists, the function creates
a backup copy in the same directory.
}
\examples{
bib <- bibentry("Misc",
  title = "My title",
  author = "Fran Pérez"
)

my_temp_bib <- tempfile(fileext = ".bib")

cff_write_bib(bib, file = my_temp_bib)

cat(readLines(my_temp_bib), sep = "\n")

cff_write_bib(bib, file = my_temp_bib, ascii = TRUE, append = TRUE)

cat(readLines(my_temp_bib), sep = "\n")
# Create an R citation file.
# Use a system file.
f <- system.file("examples/preferred-citation-book.cff", package = "cffr")
a_cff <- cff_read(f)

out <- file.path(tempdir(), "CITATION")
cff_write_citation(a_cff, file = out)

# Check by reading with a meta object.
meta <- packageDescription("cffr")
meta$Encoding <- "UTF-8"

utils::readCitationFile(out, meta)
}
\references{
R Core Team (2026). \emph{Writing R Extensions}.
\url{https://cran.r-project.org/doc/manuals/r-release/R-exts.html}.
}
\seealso{
\code{vignette("bibtex-cff", package = "cffr")}, \code{\link[knitr:write_bib]{knitr::write_bib()}} and the
following packages:
\itemize{
\item \CRANpkg{bibtex}.
\item \CRANpkg{RefManageR}.
\item \CRANpkg{rbibutils}.
}

Work with BibTeX metadata:
\code{\link[=as_bibentry]{as_bibentry()}},
\code{\link[=cff_read]{cff_read()}},
\code{\link[=cff_read_bib_text]{cff_read_bib_text()}},
\code{\link[=encoded_utf_to_latex]{encoded_utf_to_latex()}}

Write citation metadata files:
\code{\link[=cff_write]{cff_write()}}
}
\concept{bibtex}
\concept{writing}
