% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cffr-package.R
\docType{package}
\name{cffr-package}
\alias{cffr-package}
\alias{cffr}
\title{cffr: Generate Citation File Format ('cff') Metadata for R Packages}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The Citation File Format version 1.2.0 (Druskat, S. et al. (2021)
\doi{10.5281/zenodo.5171937} is a human and machine readable file format
which provides citation metadata for software. This package provides core
utilities to generate and validate this metadata.
}
\seealso{
Useful links:
\itemize{
\item \url{https://dieghernan.github.io/cffr/}
\item \url{https://github.com/ropensci/cffr}
\item Report bugs at \url{https://github.com/ropensci/cffr/issues}
}
}
\author{
\strong{Maintainer}: Diego Hernangómez (\href{https://orcid.org/0000-0001-8457-4658}{ORCID})

Other contributors:
\itemize{
\item João Martins [reviewer]
\item Scott Chamberlain [reviewer]
}
}
\keyword{internal}
