% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cffR-package.R
\docType{package}
\name{cffR-package}
\alias{cffR}
\alias{cffR-package}
\title{cffR: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dieghernan/cffR}
  \item Report bugs at \url{https://github.com/dieghernan/cffR/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID})

}
\keyword{internal}
