% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cran_to_spdx}
\alias{cran_to_spdx}
\title{Mapping between \code{License} fields and SPDX}
\format{
A data frame with 91 rows and 2 variables:
\itemize{
\item LICENSE: A valid \code{License} string on CRAN.
\item SPDX. A valid SPDX License Identifier.
}
}
\source{
\url{https://spdx.org/licenses/}
}
\usage{
cran_to_spdx
}
\description{
A dataset containing the mapping between the \code{License} strings observed
on CRAN packages and its (approximate) match on the
\href{https://spdx.org/licenses/}{SPDX License List}.
}
\examples{

data("cran_to_spdx")

head(cran_to_spdx, 20)
}
\seealso{
\emph{Writing R Extensions}, \href{https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Licensing}{Licensing section}.
}
\concept{datasets}
\keyword{datasets}
