% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{cff_extract_to_bibtex}
\alias{cff_extract_to_bibtex}
\alias{cff_to_bibtex}
\title{Previous API: Create BibTeX entries from several sources}
\usage{
cff_extract_to_bibtex(x, what = c("preferred", "references", "all"))

cff_to_bibtex(x, what = c("preferred", "references", "all"))
}
\arguments{
\item{x}{The source that would be used for generating
the \code{\link[=bibentry]{bibentry()}} object via \code{cff}. It could be:
\itemize{
\item A missing value. That would retrieve the DESCRIPTION
file on your in-development package.
\item An existing \code{\link{cff}} object,
\item Path to a CITATION.cff file (\code{"*/CITATION.cff*"}),
\item The name of an installed package (\code{"jsonlite"}), or
\item Path to a DESCRIPTION file (\code{"*/DESCRIPTION*"}).
}}

\item{what}{Fields to extract. The value could be:
\itemize{
\item \code{preferred}: This would create a single entry with the main citation
info of the package.
\item \code{references}: Extract all the entries on \code{references}.
\item \code{all}: A combination of the previous two options. This would extract
both the preferred citation info and the references.
}}
}
\value{
See \code{\link[=cff_to_bibentry]{cff_to_bibentry()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} Please use \code{\link[=cff_to_bibentry]{cff_to_bibentry()}} instead.
}
\examples{
\donttest{
# From a cff object
cff_object <- cff()

cff_object

# bibentry object
bib <- cff_to_bibentry(cff_object)
}
}
\seealso{
Other deprecated functions:
\code{\link{cff_from_bibtex}()}
}
\concept{deprecated}
\keyword{internal}
